+++ src/Events/EventLibrary.cs
@@ -126,6 +126,9 @@ namespace ChatChaos.Events
             EventRegistry.Add("random_event", "Random event", "Évènement aléatoire", () =>
                 Core.RandomEvent.Trigger());
 
+            EventRegistry.Add("start_of_day", "Start of day", "Début de journée", () =>
+                EventActions.ResetToMorning());
+
             EventRegistry.Add("power_on", "Turn power on", "Allumer courant", () =>
                 EventActions.SetFacilityPower(true));